# Use a slightly faster computation for sqrt_ratio_i.